+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Change
 
 - `LongPollingServiceContext` now generic.
+- Use channel instead of callbacks.
 
 ## [0.8.0]